move resource-explorer tests
