+++ src/app/views/sidebar/resource-explorer/ResourceExplorer.tsx
@@ -18,7 +18,7 @@ import { searchBoxStyles } from '../../../utils/searchbox.styles';
 import { translateMessage } from '../../../utils/translate-messages';
 import { classNames } from '../../classnames';
 import { sidebarStyles } from '../Sidebar.styles';
-import CommandOptions from './CommandOptions';
+import CommandOptions from './command-options/CommandOptions';
 import {
   createResourcesList, getCurrentTree,
   getResourcePaths,
@@ -144,7 +144,7 @@ const unstyledResourceExplorer = (props: any) => {
     setItems(createResourcesList(filtered.children, version));
   }
 
-  const clickLink = (ev?: React.MouseEvent<HTMLElement>, item? : INavLink) => {
+  const clickLink = (ev?: React.MouseEvent<HTMLElement>, item?: INavLink) => {
     ev!.preventDefault();
     item!.isExpanded = !item!.isExpanded;
     setQuery(item!);
